feat : PROJ-98 : 즐겨찾기 로직 구현

+++ server/Spring/src/main/java/com/hanium/diarist/domain/diary/repository/DiaryRepository.java
@@ -10,5 +10,6 @@ import java.util.Optional;
 public interface DiaryRepository extends JpaRepository<Diary, Long> {
 
     Optional<Diary> findByUserAndDiaryDate(User user, LocalDate diaryDate);
-    Optional<Diary> findByDiaryId(Integer diaryId);
+    Optional<Diary> findByDiaryId(long diaryId);
+
 }
